+++ packages/expo-router/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Upgraded React Navigation
+
 ## 5.0.1-preview.0 — 2025-04-05
 
 _This version does not introduce any user-facing changes._